requirements.txt
asgiref==3.6.0
DateTime==5.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
zope.interface==5.5.2
